move run status and shutodwn behavior to task
