+++ core/src/config/tasks.ts
@@ -11,7 +11,15 @@ export const DEFAULT = {
         const { api } = await import("actionhero"); // this needs to be async loaded as we are within the config system, to avoid circular dependencies
 
         return [].concat(
-          ["imports", "events", "runs", "schedules", "groups", "exports"],
+          [
+            "system",
+            "imports",
+            "events",
+            "runs",
+            "schedules",
+            "groups",
+            "exports",
+          ],
           api?.plugins?.plugins
             .filter((plugin) => plugin.apps?.length > 0)
             .map((plugin) => plugin.apps.map((app) => `exports:${app.name}`)),
